% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cross.val}
\alias{cross.val}
\title{multi-fold cross-validation}
\usage{
cross.val(data, na = -999, n.fold = 10L, lambda = 1, control = list())
}
\arguments{
\item{data}{Data frame. Not necessarily on the [0, 1] scale.}

\item{na}{A numeric value which is to be interpreted as NA values. The default is -999.}

\item{n.fold}{Number of folds The default is 10.}

\item{lambda}{Vector of penalty weights. The default is 1.}

\item{control}{List of control variables.}
}
\value{

}
\description{
Wrapper function to perform multi-fold cross-validation.
}
