% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DenTrans}
\alias{DenTrans}
\title{marginal probability integral transform}
\usage{
DenTrans(data, logT = T, na.strings = -999)
}
\arguments{
\item{data}{Response time data or log response time data on a scale different from 0 to 1.}

\item{logT}{a logical value indicating whether response time data is on a log scale. The default is \code{TRUE}}

\item{na.strings}{a character vector of strings which are to be interpreted as NA values. The default is -999.}
}
\value{
return response time data scaled from 0 to 1 with -999 as the missing value.
}
\description{
Transform a random variable into a scale of 0 and 1 using the marginal 
probability integral transformation
}
