% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{loglik.out}
\alias{loglik.out}
\title{Get the loglikelihood}
\usage{
loglik.out(par, By, Bx, Bq, qwt)
}
\arguments{
\item{par}{item parameters; numeric, dim = p = py + py * px}

\item{By}{basis for y}

\item{Bx}{basis for x}

\item{Bq}{basis for quadrature points}

\item{qwt}{quadrature weights}
}
\value{
log-likelihood
}
\description{
Get the loglikelihood from the fitted model to prepare for the density plots
}
