% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minmax}
\alias{minmax}
\title{min max method}
\usage{
minmax(data, na = -999)
}
\arguments{
\item{data}{Data frame. Not necessarily on the [0, 1] scale.}

\item{na}{A numeric value which is to be interpreted as NA values. The default is -999.}
}
\value{

}
\description{
To scale a manifest variable into the range from 0 to 1.
}
