% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{persp.ecdns}
\alias{persp.ecdns}
\title{perspective plot for estimated conditional density}
\usage{
\method{persp}{ecdns}(par, x = seq(0, 1, 0.02), y = seq(0, 1, 0.02), nby = 11, nbx = 11, ...)
}
\arguments{
\item{par}{item parameters; numeric, dim = p = py + py * px}

\item{x}{x values for plotting (numeric)}

\item{y}{x values for plotting (numeric)}

\item{nby}{number of bases for y}

\item{nbx}{number of bases for x}

\item{...}{additional parameters for perspective plots}
}
\value{

}
\description{
perspective plot for estimated conditional density
}
\seealso{
\code{\link{persp}},\code{\link{contour}}, and \code{\link{persp.ecdns}}
}
