% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfa_continous.R
\name{spfa_continous}
\alias{spfa_continous}
\title{Fit one-dimensional SPFA for continuous data (response time)}
\usage{
spfa_continous(RT = NULL, na.string = -999, control = list())
}
\arguments{
\item{RT}{Response time data in a matrix format.}

\item{na.string}{A numeric value which is to be interpreted as NA values. The default is -999.}

\item{control}{A list of control variables.}
}
\value{

}
\description{
Fit one-dimensional SPFA for continuous data (response time)
}
