% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfa_discrete.R
\name{spfa_discrete}
\alias{spfa_discrete}
\title{Fit one-dimensional SPFA for discrete data (item responses)}
\usage{
spfa_discrete(
  dat = NULL,
  na.string = -999,
  control = list(n_basis = 11, lmbd = 10)
)
}
\arguments{
\item{dat}{A matrix of item response data which contains \code{ncol(dat)} of discrete item responses.}

\item{na.string}{A numeric value which is to be interpreted as NA values. The default is -999.}

\item{control}{A list of control variables.}
}
\value{

}
\description{
Fit one-dimensional SPFA for discrete data (item responses)
}
