% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfa_example.R
\docType{data}
\name{spfa_example}
\alias{spfa_example}
\title{spfa_example}
\description{
An R object containing an example output from fitting an spfa model using the following code:
# RT <- spfa::simdata[,1:8]
# spfa_example <- spfa(data = RT, dimension = rep(0, ncol(RT)), discrete = rep(FALSE, ncol(RT)))
}
\keyword{data}
