% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfa-package.R
\docType{package}
\name{spfa-package}
\alias{spfa}
\alias{spfa-package}
\title{spfa: Fitting semiparametric factor analyses model that incorporates item responses and response time}
\description{
More about what it does (maybe more than one line)
}
\author{
\strong{Maintainer}: Yang Liu \email{yliu87@umd.edu}

Authors:
\itemize{
  \item Weimeng Wang \email{weimengbonnie@gmail.com} [contributor]
}

}
\keyword{internal}
