\name{spfa-package}
\alias{spfa-package}
\alias{spfa}
\docType{package}
\title{
\packageTitle{spfa}
}
\description{
\packageDescription{spfa}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spfa}
\packageIndices{spfa}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{spfa}

Maintainer: \packageMaintainer{spfa}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
